Sanitize CSVs to protect against CSV Injection attacks
A csv cell starts with `=` can be interpreted as formula[1],  which needed to be sanitized.

In addition, according to a ticket of Tabulator [2], the only way to access download type from accessorDownload is to read it from some global variables. In this case it is a private field in class.

[1] https://owasp.org/www-community/attacks/CSV_Injection
[2] https://github.com/olifolkerd/tabulator/issues/2111